+++ webapp/channels/src/types/store/plugins.ts
@@ -8,10 +8,10 @@ import type {WebSocketClient} from '@mattermost/client';
 import type {IconGlyphTypes} from '@mattermost/compass-icons/IconGlyphs';
 import type {PluginAnalyticsRow} from '@mattermost/types/admin';
 import type {Board} from '@mattermost/types/boards';
-import type {Channel, ChannelMembership, CreateResult, NewChannelFormState} from '@mattermost/types/channels';
+import type {Channel, ChannelMembership} from '@mattermost/types/channels';
 import type {FileInfo} from '@mattermost/types/files';
 import type {CommandArgs} from '@mattermost/types/integrations';
-import type {ClientPluginManifest} from '@mattermost/types/plugins';
+import type {ClientPluginManifest, CreateResult, NewChannelFormState} from '@mattermost/types/plugins';
 import type {Post, PostEmbed} from '@mattermost/types/posts';
 import type {ProductScope} from '@mattermost/types/products';
 import type {UserProfile} from '@mattermost/types/users';
