+++ webapp/platform/types/src/channels.ts
@@ -288,29 +288,3 @@ export type ChannelSearchOpts = {
     parent_access_control_policy_id?: string;
 };
 
-/**
- * The form state passed to a plugin's channel-type option when the user submits channel creation.
- * `type` identifies which option was selected so a plugin that registers multiple options can branch.
- * `teamId` and `type` are owned by the modal and are not writable via setFormState.
- */
-export type NewChannelFormState = {
-    teamId: string;
-    displayName: string;
-    url: string;
-    purpose: string;
-    type: string;
-    managedCategoryName?: string;
-};
-
-/**
- * Discriminated union returned by a plugin's `onCreate` handler.
- * - `created`: plugin created the channel successfully.
- * - `deferred`: plugin will handle creation asynchronously; the modal closes immediately.
- * - `error`: plugin encountered an error; the modal displays `message`.
- * - `cancelled`: plugin aborted creation; the modal stays open without an error.
- */
-export type CreateResult =
-    | {status: 'created'; channel: Channel}
-    | {status: 'deferred'}
-    | {status: 'error'; message: string}
-    | {status: 'cancelled'};
